Add content for session 5 (sending emails)
Although based on the Python notes, I've had to change them to using
Mailgun, as Mailchimp now requires a DNS-verified domain to send emails
(which is way out of the scope for our course). Mailgun offers a sandbox
domain, which doesn't require DNS verification, and works well for our
purposes.

Mailgun does have an API gem available, but it depends on the `unf` gem
which uses a C extension (meaning that we would need to set students up
with a C compiler,something I would be keen to avoid), and triggers a
no-SSL-certificates-available error on a stock Ruby install on Mac OS X.

We thus use the Pony gem to send emails via SMTP (as per the old notes).
Since we're no longer really using an external API, I've punted the API
explanations from the Python notes to the Twitter session.